% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2str_functions.R
\name{codes}
\alias{codes}
\title{Integer Codes of Factor Levels}
\usage{
codes(fct)
}
\arguments{
\item{fct}{factor.}
}
\value{
integer vector with length = \code{length(levels(fct))}, elements = integer
codes of \code{fct} and names = \code{levels(fct)}.
}
\description{
\code{codes} returns the integer codes for each factor level from a factor.
}
\examples{
codes(state.region)
codes(iris$"Species")
}
