% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2str_functions.R
\name{is.Date}
\alias{is.Date}
\title{Test for a Date object}
\usage{
is.Date(x)
}
\arguments{
\item{x}{an object.}
}
\value{
TRUE is \code{x} has class "Date" and FALSE if \code{x} does not have class "Date".
}
\description{
\code{is.Date} returns whether an object is a Date object (aka has class = "Date").
}
\examples{
date <- as.Date("2021-05-24", format = "\%Y-\%m-\%d") # as.Date.character
is.Date(date)
class(date) <- append(class(date), "extra_class")
is.Date(date) # classes other than Date are allowed
is.Date(list(date)) # returns FALSE
}
