% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2str_functions.R
\name{ndim}
\alias{ndim}
\title{Number of Object Dimensions}
\usage{
ndim(x)
}
\arguments{
\item{x}{object that has dimensions (e.g., array).}
}
\value{
integer vector of length 1 specifying the number of dimensions in \code{x}.
If \code{x} does not have any dimensions, then O is returned.
}
\description{
\code{ndim} returns the number of dimensions an object has. This is most useful
for arrays, which can have anywhere from 1 to 1000+ dimensions.
}
\details{
\code{ndim} is a very simple function that is simply \code{length(dim(x))}.
}
\examples{
ndim(state.region)
ndim(attitude)
ndim(HairEyeColor)
}
