% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2str_functions.R
\name{sn}
\alias{sn}
\title{Set a Vector's Names as its Elements}
\usage{
sn(x)
}
\arguments{
\item{x}{atomic or list vector.}
}
\value{
\code{x} with the elements of \code{x} as its names.
}
\description{
\code{sn} sets a vector's names as its elements. It is a simple utility function
equal to \code{setNames(x, nm = as.character(x))}. This is particularly useful
when using \code{lapply} and you want the return object to have \code{X} as its names.
}
\examples{
sn(1:10)
sn(c("one","two","three"))
}
